Move parameter names to constants
